api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "base58>=2.1.1",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.0",
    "pypolkadot>=0.2.5",
    "sqlmodel>=0.0.37",
    "substrate-interface>=1.8.0",
    "yt-dlp>=2026.3.3",
]
